improve socket 'project-user-remove' event handler

+++ backend/socketControllers/socketProjectController.js
@@ -236,33 +236,36 @@ export const socketProjectController = (io, socket) => {
           io.to(projectId).emit('tasks-updated', { tasks: updatedTasks });
         });
       }
-      projectData.users.splice(userIndex, 1);
+      const [deletedUser] = projectData.users.splice(userIndex, 1);
       io.to(projectId).emit('project-users-updated', projectData.users);
-      const user = await User.findOneAndUpdate(
-        { _id: userId },
-        {
-          $pull: { projectsJoined: projectId },
-          $unset: { [`projectsThemes.${projectId}`]: '' },
-        }
-      );
 
-      // if user did set project background then delete it from db
-      const background = user.projectsThemes[projectId].background;
-      if (!background.startsWith('linear'))
-        await deleteImage(background.split('images/')[1]);
-
-      // remove all notifications regarding this project and send notification about removal
       await Notification.deleteMany({ recipient: userId, project: projectId });
-      if (!socket.user._id.equals(userId)) {
-        const removeNotification = new Notification({
-          type: 'Removed From Project',
-          description: ` removed you from project: ${projectData.title}`,
-          project: projectId,
-          seenDate: null,
-          sender: socket.user._id,
-          recipient: userId,
-        });
-        await removeNotification.save();
+      if (deletedUser !== -1 && deletedUser.permissions > 0) {
+        const user = await User.findOneAndUpdate(
+          { _id: userId },
+          {
+            $pull: { projectsJoined: projectId },
+            $unset: { [`projectsThemes.${projectId}`]: '' },
+          }
+        );
+
+        // if user did set project background then delete it from db
+        const background = user?.projectsThemes[projectId]?.background;
+        if (background && !background.startsWith('linear'))
+          await deleteImage(background.split('images/')[1]);
+
+        // remove all notifications regarding this project and send notification about removal
+        if (!socket.user._id.equals(userId)) {
+          const removeNotification = new Notification({
+            type: 'Removed From Project',
+            description: ` removed you from project: ${projectData.title}`,
+            project: projectId,
+            seenDate: null,
+            sender: socket.user._id,
+            recipient: userId,
+          });
+          await removeNotification.save();
+        }
       }
       io.to(userId).emit('notifications-updated');
     }
